+++ CHANGELOG.md
@@ -6,3 +6,7 @@ details, see the commit logs. The format is based on
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+
+- Virshx script